+++ core/src/routes/(modules)/options/common/background-attachment.svelte
@@ -1,0 +1,116 @@
+<script>
+
+    import "../../../../app.css";
+    import Select from "../../../uicomponents/select.svelte";
+
+    /**
+     * @type boolean
+     */
+    export let loaded;
+
+    /**
+     * Actual class
+     * @type string
+     */
+    export let classInput;
+
+    /**
+     * Actual part of class
+     * @type string
+     */
+    export let property;
+
+
+    /**
+     * @typedef {Object} SelectOptions
+     * @property {string} value
+     * @property {string} name
+     * @property {string} info
+    */
+
+
+    /**
+     * Value of element
+     * @type Array<SelectOptions>
+     */
+    let propertyOptions = [
+        
+        {name: "Default",  value: "",               info: "Default"},
+        
+        {name: "Fixed",     value: "bg-fixed",     info: "background-attachment: fixed;\n\nUse 'Fixed' to fix the background image relative to the viewport."},
+        {name: "Local",     value: "bg-local",     info: "background-attachment: local;\n\nUse 'Local' to scroll the background image with the container and the viewport."},
+        {name: "Scroll",    value: "bg-scroll",    info: "background-attachment: scroll;\n\nUse 'Scroll' to scroll the background image with the viewport, but not with the container."},
+
+        
+
+        // {name: "",     value: "",     info: ""},
+       
+    ];
+
+    /**
+     * List of all classes
+     * @type Array<string>
+     */
+    let propertyList = [];
+
+    propertyOptions.forEach(elm => {
+        if(elm.value != "") propertyList.push(elm.value);
+    });
+
+    import { createEventDispatcher } from 'svelte';
+
+    const dispatch = createEventDispatcher();
+
+    function updateClass() {
+        dispatch('updateClass');
+    }
+
+    /**
+     * Used for sub elements update control.
+     * @type string
+    */
+    export let elementDataLoaded;
+
+    $: loaded, (()=>{
+        if(loaded == true) {
+            elementDataLoaded.split(" ").forEach( cls => {
+
+                var currentClass = cls.trim();
+
+                propertyList.forEach( elm => {
+                    if(elm === currentClass) property = elm;
+                });
+
+
+            });
+        }
+    })();
+
+    $: property, (()=>{
+        if(loaded == true) {
+
+            var newClass = "";
+
+            classInput.split(" ").forEach( cls => {
+
+                if(propertyList.includes(cls.trim())){
+                    ///Do nothing - If included then should be removed
+                }else{
+                    newClass += " " + cls.trim();
+                }
+            });
+
+            classInput = newClass.trim() + " " + property;
+            updateClass();
+        }
+    })();
+
+
+
+</script>
+
+
+<div class="w-full flex flex-row flex-grow justify-between h-8 align-middle items-center">
+    <span title="Utilities for controlling how a background image behaves when scrolling.">Attachment</span>
+    <Select options={propertyOptions} bind:value={property}/>
+</div>